+++ tests/test_live.py
@@ -1,13 +1,10 @@
 import os
 
 import pytest
-import yaml
 
 from subsetter.common import DatabaseConfig
-from subsetter.planner import Planner, PlannerConfig, SubsetPlan
-from subsetter.sampler import DatabaseOutputConfig, Sampler, SamplerConfig
 
-from .dataset_manager import apply_dataset, get_rows
+from .dataset_manager import do_dataset_test
 
 DATASETS_BASE_PATH = os.path.join(os.path.dirname(__file__), "datasets")
 
@@ -60,57 +57,10 @@ def fixture_db_config(request):
 
 
 @pytest.mark.parametrize("db_config", DATABASE_CONFIGURATIONS, indirect=True)
-def test_basic_subset(db_config):
-    apply_dataset(db_config, "user_orders")
-    config = PlannerConfig(
-        source=db_config,
-        targets={
-            "test.users": PlannerConfig.TargetConfig(  # type: ignore
-                **{"in": {"sample": [1]}}
-            ),
-        },
-        select=[
-            "test.*",
-        ],
-    )
-    planner = Planner(config)
-    plan = planner.plan()
+def test_user_orders(db_config):
+    do_dataset_test(db_config, "user_orders")
 
-    expected_plan_file = os.path.join(
-        os.path.dirname(__file__), "basic_plan_expected.yml"
-    )
-    with open(expected_plan_file, "r", encoding="utf-8") as fexpected:
-        expected_plan = SubsetPlan.model_validate(yaml.safe_load(fexpected))
 
-    assert plan == expected_plan
-
-    sample_config = SamplerConfig(
-        source=db_config,
-        output=DatabaseOutputConfig(
-            mode="database",
-            remap=[
-                {
-                    "search": r"^(\w+)\.",
-                    "replace": r"\1_out.",
-                },
-            ],
-            **db_config.model_dump(),
-        ),
-    )
-
-    sampler = Sampler(sample_config)
-    sampler.sample(plan)
-
-    sample = {
-        "users": get_rows(db_config, "test_out", "users"),
-        "orders": get_rows(db_config, "test_out", "orders"),
-        "order_status": get_rows(db_config, "test_out", "order_status"),
-    }
-
-    expected_sample_file = os.path.join(
-        os.path.dirname(__file__), "basic_sample_expected.yml"
-    )
-    with open(expected_sample_file, "r", encoding="utf-8") as fexpected:
-        expected_sample = yaml.safe_load(fexpected)
-
-    assert sample == expected_sample
+@pytest.mark.parametrize("db_config", DATABASE_CONFIGURATIONS, indirect=True)
+def test_data_types(db_config):
+    do_dataset_test(db_config, "data_types")